+++ src/sound/al_Biquad.cpp
@@ -6,9 +6,9 @@
 //
 //
 #include "al/sound/al_Biquad.hpp"
-#include <stdlib.h>
-#include <cmath>
 #include "al/math/al_Constants.hpp"
+#include <cmath>
+#include <stdlib.h>
 
 using namespace al;
 
@@ -24,79 +24,81 @@ BiQuad::~BiQuad() {}
 
 void BiQuad::set(double freq, double bandwidth, double dbGain) {
   // TODO all the way to fs/2, range
-  if (freq > 20000) freq = 20000;
-  if (freq <= 20) freq = 20;
+  if (freq > 20000)
+    freq = 20000;
+  if (freq <= 20)
+    freq = 20;
 
   double A, omega, sn, cs, alpha, beta;
   double a0, a1, a2, b0, b1, b2;
 
   // setup variables
   A = pow(10, dbGain / 40);
-  omega = 2 * M_PI * freq / (1 * mSampleRate);  // 1X or 2X oversampled
+  omega = 2 * M_PI * freq / (1 * mSampleRate); // 1X or 2X oversampled
   sn = sin(omega);
   cs = cos(omega);
   alpha = sn * sinh(M_LN2 / 2 * bandwidth * omega / sn);
   beta = sqrt(A + A);
 
   switch (mType) {
-    case BIQUAD_LPF:
-      b0 = (1 - cs) / 2;
-      b1 = 1 - cs;
-      b2 = (1 - cs) / 2;
-      a0 = 1 + alpha;
-      a1 = -2 * cs;
-      a2 = 1 - alpha;
-      break;
-    case BIQUAD_HPF:
-      b0 = (1 + cs) / 2;
-      b1 = -(1 + cs);
-      b2 = (1 + cs) / 2;
-      a0 = 1 + alpha;
-      a1 = -2 * cs;
-      a2 = 1 - alpha;
-      break;
-    case BIQUAD_BPF:
-      b0 = alpha;
-      b1 = 0;
-      b2 = -alpha;
-      a0 = 1 + alpha;
-      a1 = -2 * cs;
-      a2 = 1 - alpha;
-      break;
-    case BIQUAD_NOTCH:
-      b0 = 1;
-      b1 = -2 * cs;
-      b2 = 1;
-      a0 = 1 + alpha;
-      a1 = -2 * cs;
-      a2 = 1 - alpha;
-      break;
-    case BIQUAD_PEQ:
-      b0 = 1 + (alpha * A);
-      b1 = -2 * cs;
-      b2 = 1 - (alpha * A);
-      a0 = 1 + (alpha / A);
-      a1 = -2 * cs;
-      a2 = 1 - (alpha / A);
-      break;
-    case BIQUAD_LSH:
-      b0 = A * ((A + 1) - (A - 1) * cs + beta * sn);
-      b1 = 2 * A * ((A - 1) - (A + 1) * cs);
-      b2 = A * ((A + 1) - (A - 1) * cs - beta * sn);
-      a0 = (A + 1) + (A - 1) * cs + beta * sn;
-      a1 = -2 * ((A - 1) + (A + 1) * cs);
-      a2 = (A + 1) + (A - 1) * cs - beta * sn;
-      break;
-    case BIQUAD_HSH:
-      b0 = A * ((A + 1) + (A - 1) * cs + beta * sn);
-      b1 = -2 * A * ((A - 1) + (A + 1) * cs);
-      b2 = A * ((A + 1) + (A - 1) * cs - beta * sn);
-      a0 = (A + 1) - (A - 1) * cs + beta * sn;
-      a1 = 2 * ((A - 1) - (A + 1) * cs);
-      a2 = (A + 1) - (A - 1) * cs - beta * sn;
-      break;
-    default:
-      return;
+  case BIQUAD_LPF:
+    b0 = (1 - cs) / 2;
+    b1 = 1 - cs;
+    b2 = (1 - cs) / 2;
+    a0 = 1 + alpha;
+    a1 = -2 * cs;
+    a2 = 1 - alpha;
+    break;
+  case BIQUAD_HPF:
+    b0 = (1 + cs) / 2;
+    b1 = -(1 + cs);
+    b2 = (1 + cs) / 2;
+    a0 = 1 + alpha;
+    a1 = -2 * cs;
+    a2 = 1 - alpha;
+    break;
+  case BIQUAD_BPF:
+    b0 = alpha;
+    b1 = 0;
+    b2 = -alpha;
+    a0 = 1 + alpha;
+    a1 = -2 * cs;
+    a2 = 1 - alpha;
+    break;
+  case BIQUAD_NOTCH:
+    b0 = 1;
+    b1 = -2 * cs;
+    b2 = 1;
+    a0 = 1 + alpha;
+    a1 = -2 * cs;
+    a2 = 1 - alpha;
+    break;
+  case BIQUAD_PEQ:
+    b0 = 1 + (alpha * A);
+    b1 = -2 * cs;
+    b2 = 1 - (alpha * A);
+    a0 = 1 + (alpha / A);
+    a1 = -2 * cs;
+    a2 = 1 - (alpha / A);
+    break;
+  case BIQUAD_LSH:
+    b0 = A * ((A + 1) - (A - 1) * cs + beta * sn);
+    b1 = 2 * A * ((A - 1) - (A + 1) * cs);
+    b2 = A * ((A + 1) - (A - 1) * cs - beta * sn);
+    a0 = (A + 1) + (A - 1) * cs + beta * sn;
+    a1 = -2 * ((A - 1) + (A + 1) * cs);
+    a2 = (A + 1) + (A - 1) * cs - beta * sn;
+    break;
+  case BIQUAD_HSH:
+    b0 = A * ((A + 1) + (A - 1) * cs + beta * sn);
+    b1 = -2 * A * ((A - 1) + (A + 1) * cs);
+    b2 = A * ((A + 1) + (A - 1) * cs - beta * sn);
+    a0 = (A + 1) - (A - 1) * cs + beta * sn;
+    a1 = 2 * ((A - 1) - (A + 1) * cs);
+    a2 = (A + 1) - (A - 1) * cs - beta * sn;
+    break;
+  default:
+    return;
   }
 
   mBD.a0 = b0 / a0;
@@ -108,7 +110,7 @@ void BiQuad::set(double freq, double bandwidth, double dbGain) {
 
 void BiQuad::processBuffer(float *buffer, int count) {
   for (int i = 0; i < count; i++) {
-    double sample = buffer[i];
+    float sample = buffer[i];
     sample = (*this)(sample);
     buffer[i] = sample;
   }
@@ -117,7 +119,8 @@ void BiQuad::processBuffer(float *buffer, int count) {
 double BiQuad::operator()(double sample) {
   double result;
 
-  if (!enabled) return sample;
+  if (!enabled)
+    return sample;
 
   // compute result
   result = mBD.a0 * sample + mBD.a1 * mBD.x1 + mBD.a2 * mBD.x2 -
@@ -154,18 +157,22 @@ double BiQuad::operator()(double sample) {
 BiQuadNX::BiQuadNX(int _numFilters, BIQUADTYPE _type, double _sampleRate)
     : numFilters(_numFilters) {
   mFilters = (BiQuad *)malloc(sizeof(BiQuad) * numFilters);
-  for (int i = 0; i < numFilters; i++) mFilters[i] = BiQuad(_type, _sampleRate);
+  for (int i = 0; i < numFilters; i++)
+    mFilters[i] = BiQuad(_type, _sampleRate);
 }
 BiQuadNX::~BiQuadNX() { free(mFilters); }
 
 void BiQuadNX::set(double freq, double bandwidth, double dbGain) {
-  for (int i = 0; i < numFilters; i++) mFilters[i].set(freq, bandwidth, dbGain);
+  for (int i = 0; i < numFilters; i++)
+    mFilters[i].set(freq, bandwidth, dbGain);
 }
 void BiQuadNX::setSampleRate(double _rate) {
-  for (int i = 0; i < numFilters; i++) mFilters[i].setSampleRate(_rate);
+  for (int i = 0; i < numFilters; i++)
+    mFilters[i].setSampleRate(_rate);
 }
 void BiQuadNX::processBuffer(float *buffer, int count) {
-  for (int i = 0; i < numFilters; i++) mFilters[i].processBuffer(buffer, count);
+  for (int i = 0; i < numFilters; i++)
+    mFilters[i].processBuffer(buffer, count);
 }
 
 double BiQuadNX::operator()(double sample) {
@@ -178,5 +185,6 @@ double BiQuadNX::operator()(double sample) {
 }
 
 void BiQuadNX::enable(bool on) {
-  for (int i = 0; i < numFilters; i++) mFilters[i].enable(on);
+  for (int i = 0; i < numFilters; i++)
+    mFilters[i].enable(on);
 }
